Inline byvalue_scanner.
